bot/pyproject.toml
[project]
name = "radioplatform-bot"
version = "0.1.0"
description = "Telegram bot for radioplatform — Claude Code over Telegram"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "openai>=1.50.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
